requirements.txt
fastapi
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings
uvicorn
sqlalchemy